Cave revisions and bugfixes
	-Altered torch light color so it looks more torch-y

	-Conditionals change movement commands from "run" to
	 "jump" when in cave sections.

	-Fixed a longstanding loss-of-precision bug in the
	 terrain generation routines that was causing blank
	 tiles and strange behavior in the world map.

	-Fixed a stupid "bug" that was preventing the inventory
	 menu from being properly initialized.

+++ src/map/terrain.h
@@ -31,10 +31,10 @@
 //#define  ELEV(val,n,...) or_byte(val, ALT, n, __VA_ARGS__)
 
 static const int   OCTAVES =  6;      /* Number of octaves to smooth */
-static const float PERSIST =  0.99;   /* Persistence factor */
-static const float SHOAL   = -0.003;  /* Shoal isoline */
-static const float BEACH   =  0.00;   /* Beach isoline */
-static const float TERRA   =  0.45;   /* Ground isoline */
+static const double PERSIST =  0.99;   /* Persistence factor */
+static const double SHOAL   = -0.003;  /* Shoal isoline */
+static const double BEACH   =  0.00;   /* Beach isoline */
+static const double TERRA   =  0.45;   /* Ground isoline */
 
 struct map_t;
